requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
openai
